change class hierarchy of RouteAdmin and MinimalRouteAdmin
